updated main.java
Added an encode() method that receives the root and inputted text, iterates through the text and pulls out the letters, then calls a new method, encodeLetter(), that pieces together the morse code of the text. The encodeLetter() method also accounts for if there is a space in the string and if the nodes are null.

I added a new scanner object in the main() method called scanner where the user enters their text, which is stored in a string called StringToConvert and passes it to the encode method.